Added support for reading from a snapsot
If bindSelf is true (default) and the window is also bound to the context group then it will register the selected context group in it's options under interop. If bindSelf is false then the window/view will never assign itself so we store it under customData so that it can still be used on load.