Added a react implementation of the context group picker and a react basic platform sample
Also fixed a bug in the context group picker where if a view with a context group was dragged into a window that had multiple views but no context group set then it would send the window context group and the context for the dragged in view but would not update any other view.

This is an initial push and more documentation will be added.